Update infoFinder.py
The infoFinder will find information on the subjects based on the occurence of the different words in the search field. If, for example someone searched "ninja robot cool" all of those words would need to be present in the paragraph for it to be returned.

Also, the quotes can be used to specify a set of words to be read litterally, for example: 'my "fur coat" ' will search if my and "fur coat" are in the paragraph, as is.